requirements.txt
absl-py==0.13.0
certifi==2021.5.30
cycler==0.10.0
einops==0.3.0
fancycompleter==0.9.1
flatbuffers==2.0
jax==0.2.20
jaxlib==0.1.71
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
opt-einsum==3.3.0
pdbpp==0.10.3
Pillow==8.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyrepl==0.9.0
python-dateutil==2.8.2
scipy==1.7.1
six==1.16.0
wmctrl==0.4
